# packages/contracts/src/generated/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/session": {
      "post": {
        "operationId": "SessionController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/SignInRequestDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SignInResponseDto_Output" }
              }
            }
          },
          "400": {
            "description": "Malformed body (`validation`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          },
          "401": {
            "description": "Adapter rejected the credential (`invalid-credentials`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          },
          "403": {
            "description": "User exists but is disabled (`user-disabled`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          },
          "429": {
            "description": "Authentication attempts are rate limited.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          }
        },
        "tags": ["Session"]
      },
      "delete": {
        "operationId": "SessionController_signOut",
        "parameters": [],
        "responses": { "204": { "description": "" } },
        "tags": ["Session"]
      }
    },
    "/v1/me": {
      "get": {
        "operationId": "MeController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MeResponseDto_Output" }
              }
            }
          },
          "401": {
            "description": "No/invalid session (`session-invalid`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          }
        },
        "tags": ["Me"]
      }
    },
...
    "/v1/invitations": {
      "post": {
        "operationId": "InvitationsController_issue",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/IssueInvitationRequestDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/IssueInvitationResponseDto_Output" }
              }
            }
          },
          "400": {
            "description": "Bad email/role, or `role: owner` (`validation`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          },
          "401": {
            "description": "No/invalid session (`session-invalid`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          },
          "403": {
            "description": "Missing `members:invite`, or no active workspace (`forbidden`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          },
          "409": {
            "description": "A pending invitation already exists for that email (`invitation-exists`), or the email already maps to an active membership (`already-member`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          }
        },
        "tags": ["Invitations"]
      }
    },
    "/v1/invitations/{token}": {
      "get": {
        "operationId": "InvitationsController_preview",
        "parameters": [
          { "name": "token", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InvitationPreviewResponseDto_Output" }
              }
            }
          },
          "404": {
            "description": "Unknown/garbage/revoked token (`invitation-not-found`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          },
          "410": {
            "description": "Expired or already-used token (`invitation-expired`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          },
          "429": {
            "description": "Too many requests for this token/IP (`rate-limited`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          }
        },
        "tags": ["Invitations"]
      }
    },
...
    "/v1/invitations/{id}": {
      "patch": {
        "operationId": "InvitationsController_revoke",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RevokeInvitationRequestDto" }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Malformed id/body (`validation`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          },
          "401": {
            "description": "No/invalid session (`session-invalid`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          },
          "403": {
            "description": "Missing `members:invite`, or no active workspace (`forbidden`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          },
          "404": {
            "description": "No pending invitation with that id in the active workspace (`not-found`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          }
        },
        "tags": ["Invitations"]
      }
    },
    "/v1/catalog/services": {
      "get": {
        "operationId": "ServicesController_list",
        "parameters": [
          { "name": "limit", "required": false, "in": "query", "schema": { "type": "integer" } },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": { "format": "uuid", "type": "string" }
          },
          {
            "name": "active",
            "required": false,
            "in": "query",
            "schema": { "enum": ["true", "false"], "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServiceListResponseDto_Output" }
              }
            }
          },
          "400": {
            "description": "Malformed query parameter (`validation`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          },
          "401": {
            "description": "No/invalid session (`session-invalid`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          },
          "403": {
            "description": "Missing `catalog:read`, or no active workspace (`forbidden`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          }
        },
        "tags": ["Services"]
      },
      "post": {
        "operationId": "ServicesController_create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Client-supplied replay key. Retrying with the same key and the same body returns the original 201 response and creates no second service; the same key with a materially different body is a 409 conflict.",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateServiceRequestDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServiceResponseDto_Output" }
              }
            }
          },
          "400": {
            "description": "Malformed body, unknown property, or missing `Idempotency-Key` (`validation`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          },
          "401": {
            "description": "No/invalid session (`session-invalid`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          },
          "403": {
            "description": "Missing `catalog:manage`, or no active workspace (`forbidden`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          },
          "409": {
            "description": "`Idempotency-Key` reused with a different request (`idempotency-conflict`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          },
          "422": {
            "description": "Duration/buffer/price/currency invariant violated, or an unavailable category (`validation`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          }
        },
        "tags": ["Services"]
      }
    },
    "/v1/catalog/services/{id}": {
      "get": {
        "operationId": "ServicesController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServiceResponseDto_Output" }
              }
            }
          },
          "401": {
            "description": "No/invalid session (`session-invalid`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          },
          "403": {
            "description": "Missing `catalog:read`, or no active workspace (`forbidden`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          },
          "404": {
            "description": "No such service in the active workspace (`not-found`). A service belonging to another workspace is indistinguishable from one that does not exist.",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          }
        },
        "tags": ["Services"]
      },
      "patch": {
        "operationId": "ServicesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateServiceRequestDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServiceResponseDto_Output" }
              }
            }
          },
          "400": {
            "description": "Malformed body or unknown property (`validation`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          },
          "401": {
            "description": "No/invalid session (`session-invalid`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          },
          "403": {
            "description": "Missing `catalog:manage`, or no active workspace (`forbidden`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          },
          "404": {
            "description": "No such service in the active workspace (`not-found`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          },
          "422": {
            "description": "Duration/buffer/price/currency invariant violated, or an unavailable category (`validation`).",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetailsDto" }
              }
            }
          }
        },
        "tags": ["Services"]
      }
    },
...
  },
  "info": {
    "title": "Slotnova API",
    "description": "Slotnova platform-foundation HTTP contract (Phase 1).",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "SignInRequestDto": {
        "type": "object",
        "properties": { "credential": { "type": "object", "additionalProperties": {} } },
        "required": ["credential"]
      },
      "SignInResponseDto_Output": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "id": { "type": "string" },
              "displayName": { "type": "string" },
              "email": { "type": "string" }
            },
            "required": ["id", "displayName", "email"],
            "additionalProperties": false
          },
          "activeWorkspace": {
            "type": "object",
            "properties": {
              "id": { "type": "string" },
              "name": { "type": "string" },
              "role": { "type": "string" }
            },
            "required": ["id", "name", "role"],
            "additionalProperties": false,
            "nullable": true
          },
          "workspaces": {
            "type": "array",
            "readOnly": true,
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string" },
                "name": { "type": "string" },
                "role": { "type": "string" }
              },
              "required": ["id", "name", "role"],
              "additionalProperties": false
            }
          }
        },
        "required": ["user", "activeWorkspace", "workspaces"],
        "additionalProperties": false
      },
      "MeResponseDto_Output": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "id": { "type": "string" },
              "displayName": { "type": "string" },
              "email": { "type": "string" }
            },
            "required": ["id", "displayName", "email"],
            "additionalProperties": false
          },
          "activeWorkspace": {
            "type": "object",
            "properties": {
              "id": { "type": "string" },
              "name": { "type": "string" },
              "role": { "type": "string" },
              "permissions": { "readOnly": true, "type": "array", "items": { "type": "string" } }
            },
            "required": ["id", "name", "role", "permissions"],
            "additionalProperties": false,
            "nullable": true
          },
          "workspaces": {
            "type": "array",
            "readOnly": true,
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string" },
                "name": { "type": "string" },
                "role": { "type": "string" }
              },
              "required": ["id", "name", "role"],
              "additionalProperties": false
            }
          },
          "session": {
            "type": "object",
            "properties": { "expiresAt": { "type": "string" } },
            "required": ["expiresAt"],
            "additionalProperties": false
          }
        },
        "required": ["user", "activeWorkspace", "workspaces", "session"],
        "additionalProperties": false
      },
...
      "IssueInvitationRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 320,
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "role": { "type": "string", "enum": ["admin", "manager", "staff"] }
        },
        "required": ["email", "role"]
      },
      "IssueInvitationResponseDto_Output": {
        "type": "object",
        "properties": {
          "invitation": {
            "type": "object",
            "properties": {
              "id": { "type": "string" },
              "email": { "type": "string" },
              "role": { "type": "string" },
              "status": { "type": "string" },
              "expiresAt": { "type": "string" }
            },
            "required": ["id", "email", "role", "status", "expiresAt"],
            "additionalProperties": false
          },
          "token": { "type": "string" }
        },
        "required": ["invitation"],
        "additionalProperties": false
      },
      "InvitationPreviewResponseDto_Output": {
        "type": "object",
        "properties": {
          "workspaceName": { "type": "string" },
          "role": { "type": "string" },
          "email": { "type": "string" },
          "status": { "type": "string" },
          "expiresAt": { "type": "string" }
        },
        "required": ["workspaceName", "role", "email", "status", "expiresAt"],
        "additionalProperties": false
      },
      "RevokeInvitationRequestDto": {
        "type": "object",
        "properties": { "status": { "type": "string", "enum": ["revoked"] } },
        "required": ["status"]
      },
...
      "ServiceResponseDto_Output": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "categoryId": { "type": "array", "items": { "type": "string" } },
          "durationMinutes": { "type": "number" },
          "preBufferMinutes": { "type": "number" },
          "postBufferMinutes": { "type": "number" },
          "priceAmountMinor": { "type": "number" },
          "priceCurrency": { "type": "string" },
          "active": { "type": "boolean" }
        },
        "required": [
          "id",
          "name",
          "categoryId",
          "durationMinutes",
          "preBufferMinutes",
          "postBufferMinutes",
          "priceAmountMinor",
          "priceCurrency",
          "active"
        ],
        "additionalProperties": false
      },
...
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": { "type": "string" },
          "title": { "type": "string" },
          "status": { "type": "number" },
          "detail": { "type": "string" },
          "instance": { "type": "string" },
          "errors": {
            "readOnly": true,
            "type": "array",
            "items": {
              "type": "object",
              "properties": { "path": { "type": "string" }, "message": { "type": "string" } },
              "required": ["path", "message"]
            }
          },
          "requiredCapability": { "type": "string" },
          "checks": { "type": "object", "additionalProperties": { "type": "string" } }
        },
        "required": ["type", "title", "status"]
      }
    }
  }
}